Rewrite it in Rust.
This commit changes the dashboard from a script that runs on the router
to a program that runs on the client. In this way, the program is
able to offload JSON parsing, string slicing and other work from the router
to itself, reducing the router's CPU usage greatly.

By being written in Rust, as opposed to a shell or awk script, the program
is also able to use the libssh2 library to send multiple commands over
the same SSH connection, which would be difficult for a script that uses
the `ssh` CLI to do.

The downside is that it is now a compiled program, compared to when it was just
a single .awk that needed to be scp'd to the router.

Other changes:

- Configuration is now read from a config file.

- The dashboard can now monitor the firewall logs of multiple WAN interfaces.

- The dashboard can now monitor the dpinger statistics of multiple gateways.

- More `sysctl` invocations use the binary format instead of parsing strings.
  This also allows them to be batched more easily.

- The temperatures list now shows disk temperatures from SMART info.

- The formatting of the interfaces list and firewall logs is more compact,
  and works better with IPv6.

- Fix double-counting of interface bandwidth by only counting
  AF_LINK bandwidth.

- The parser for the firewall logs is now more robust.

- Firewall logs are now read from a single long-running `clog -f` command
  instead of repeatedly spawning a new `clog | tail` command for every redraw.
  This reduces the work the router has to do, at the cost of requiring
  a second SSH connection just for this.

- Commands are now invoked with absolute paths of the binaries.